+++ release-raccoon-app/src/main/java/com/raccoon/configuration/RaccoonConfig.java
@@ -1,0 +1,10 @@
+package com.raccoon.configuration;
+
+import io.smallrye.config.ConfigMapping;
+
+@ConfigMapping(prefix = "raccoon")
+public interface RaccoonConfig {
+
+    String baseUrl();
+
+}